+++ configuration/jms-setup-wildfly.cli
@@ -87,7 +87,7 @@ catch
   /subsystem=messaging-activemq/server=default/jms-queue=FreeServiceSlotQueue:remove
 
   /subsystem=messaging-activemq/server=default/jms-topic=MetricsProcessingStartedTopic:remove
-  #/subsystem=messaging-activemq/server=default/jms-topic=APNSClient:remove
+  /subsystem=messaging-activemq/server=default/jms-topic=APNSClient:remove
 
   /subsystem=messaging-activemq/server=default/jms-queue=APNsPushMessageQueue:add(entries=[queue/APNsPushMessageQueue])
   /subsystem=messaging-activemq/server=default/address-setting=jms.queue.APNsPushMessageQueue:add(redelivery-delay=1500, redelivery-multiplier=1.5, max-redelivery-delay=5000, max-delivery-attempts=-1)
@@ -132,7 +132,7 @@ catch
 
 
   /subsystem=messaging-activemq/server=default/jms-topic=MetricsProcessingStartedTopic:add(entries=[topic/MetricsProcessingStartedTopic])
-  #/subsystem=messaging-activemq/server=default/jms-topic=APNSClient:add(entries=[topic/APNSClient])
+  /subsystem=messaging-activemq/server=default/jms-topic=APNSClient:add(entries=[topic/APNSClient])
 
 finally
 end-try
