+++ push-sender/src/main/java/org/jboss/aerogear/unifiedpush/message/jms/MessageHolderWithVariantsProducer.java
@@ -1,13 +1,13 @@
 /**
  * JBoss, Home of Professional Open Source
  * Copyright Red Hat, Inc., and individual contributors.
- *
+ * <p>
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -34,17 +34,17 @@ public class MessageHolderWithVariantsProducer extends AbstractJMSMessageProduce
 
     private static final Logger logger = LoggerFactory.getLogger(MessageHolderWithVariantsProducer.class);
 
-    
-    private  final String apnsPushMessageQueue = "APNsPushMessageQueue";
 
-    
-    private final String gcmPushMessageQueue = "GCMPushMessageQueue";
+    private static final String apnsPushMessageQueue = "APNsPushMessageQueue";
 
-    
-    private final String wnsPushMessageQueue = "WNSPushMessageQueue";
+
+    private static final String gcmPushMessageQueue = "GCMPushMessageQueue";
+
+
+    private static final String wnsPushMessageQueue = "WNSPushMessageQueue";
 
     public void queueMessageVariantForProcessing(@Observes @DispatchToQueue MessageHolderWithVariants msg) {
-        logger.debug("dispatching for processing variants and trigger token querying/batching");
+        logger.trace("dispatching for processing variants and trigger token querying/batching");
         sendTransacted(selectQueue(msg.getVariantType()), msg, false);
     }
