+++ servers/keycloak/src/main/docker/standalone-full.xml
@@ -429,7 +429,7 @@
         </subsystem>
         <subsystem xmlns="urn:jboss:domain:jaxrs:1.0"/>
         <subsystem xmlns="urn:jboss:domain:jca:5.0">
-            <org.jboss.aerogear.unifiedpush.test.archive-validation enabled="true" fail-on-error="true" fail-on-warn="false"/>
+            <archive-validation enabled="true" fail-on-error="true" fail-on-warn="false"/>
             <bean-validation enabled="true"/>
             <default-workmanager>
                 <short-running-threads>
