+++ service/pom.xml
@@ -165,7 +165,7 @@
                 <dependency>
                     <groupId>org.jboss.arquillian.container</groupId>
                     <artifactId>arquillian-container-spi</artifactId>
-                    <version>1.4.1.Final</version>
+                    <version>${arquillian.version}</version>
                     <scope>test</scope>
                 </dependency>
                 <dependency>
@@ -178,96 +178,96 @@
                 <dependency>
                     <groupId>org.jboss.arquillian.core</groupId>
                     <artifactId>arquillian-core-spi</artifactId>
-                    <version>1.4.1.Final</version>
+                    <version>${arquillian.version}</version>
                     <scope>test</scope>
                 </dependency>
                 <dependency>
                     <groupId>org.jboss.arquillian.container</groupId>
                     <artifactId>arquillian-container-test-spi</artifactId>
-                    <version>1.4.1.Final</version>
+                    <version>${arquillian.version}</version>
                     <scope>test</scope>
                 </dependency>
                 <dependency>
                     <groupId>org.jboss.arquillian.container</groupId>
                     <artifactId>arquillian-container-impl-base</artifactId>
-                    <version>1.4.1.Final</version>
+                    <version>${arquillian.version}</version>
                     <scope>test</scope>
                 </dependency>
                 <dependency>
                     <groupId>org.jboss.arquillian.container</groupId>
                     <artifactId>arquillian-container-test-impl-base</artifactId>
-                    <version>1.4.1.Final</version>
+                    <version>${arquillian.version}</version>
                     <scope>test</scope>
                 </dependency>
                 <dependency>
                     <groupId>org.jboss.arquillian.protocol</groupId>
                     <artifactId>arquillian-protocol-servlet</artifactId>
-                    <version>1.4.1.Final</version>
+                    <version>${arquillian.version}</version>
                     <scope>test</scope>
                 </dependency>
                 <dependency>
                     <groupId>org.jboss.arquillian.testenricher</groupId>
                     <artifactId>arquillian-testenricher-ejb</artifactId>
-                    <version>1.4.1.Final</version>
+                    <version>${arquillian.version}</version>
                     <scope>test</scope>
                 </dependency>
                 <dependency>
                     <groupId>org.jboss.arquillian.testenricher</groupId>
                     <artifactId>arquillian-testenricher-resource</artifactId>
-                    <version>1.4.1.Final</version>
+                    <version>${arquillian.version}</version>
                     <scope>test</scope>
                 </dependency>
                 <dependency>
                     <groupId>org.jboss.arquillian.testenricher</groupId>
                     <artifactId>arquillian-testenricher-cdi</artifactId>
-                    <version>1.4.1.Final</version>
+                    <version>${arquillian.version}</version>
                     <scope>test</scope>
                 </dependency>
                 <dependency>
                     <groupId>org.jboss.arquillian.testenricher</groupId>
                     <artifactId>arquillian-testenricher-initialcontext</artifactId>
-                    <version>1.4.1.Final</version>
+                    <version>${arquillian.version}</version>
                     <scope>test</scope>
                 </dependency>
 
                 <dependency>
                     <groupId>org.jboss.arquillian.config</groupId>
                     <artifactId>arquillian-config-api</artifactId>
-                    <version>1.4.1.Final</version>
+                    <version>${arquillian.version}</version>
                     <scope>test</scope>
                 </dependency>
 
                 <dependency>
                     <groupId>org.jboss.arquillian.core</groupId>
                     <artifactId>arquillian-core-api</artifactId>
-                    <version>1.4.1.Final</version>
+                    <version>${arquillian.version}</version>
                     <scope>test</scope>
                 </dependency>
 
                 <dependency>
                     <groupId>org.jboss.arquillian.test</groupId>
                     <artifactId>arquillian-test-api</artifactId>
-                    <version>1.4.1.Final</version>
+                    <version>${arquillian.version}</version>
                     <scope>test</scope>
                 </dependency>
                 <dependency>
                     <groupId>org.jboss.arquillian.config</groupId>
                     <artifactId>arquillian-config-spi</artifactId>
-                    <version>1.4.1.Final</version>
+                    <version>${arquillian.version}</version>
                     <scope>test</scope>
                 </dependency>
 
                 <dependency>
                     <groupId>org.jboss.arquillian.test</groupId>
                     <artifactId>arquillian-test-spi</artifactId>
-                    <version>1.4.1.Final</version>
+                    <version>${arquillian.version}</version>
                     <scope>test</scope>
                 </dependency>
 
                 <dependency>
                     <groupId>org.jboss.arquillian.config</groupId>
                     <artifactId>arquillian-config-impl-base</artifactId>
-                    <version>1.4.1.Final</version>
+                    <version>${arquillian.version}</version>
                     <scope>test</scope>
                 </dependency>
 
@@ -280,14 +280,14 @@
                 <dependency>
                     <groupId>org.jboss.arquillian.container</groupId>
                     <artifactId>arquillian-container-test-api</artifactId>
-                    <version>1.4.1.Final</version>
+                    <version>${arquillian.version}</version>
                     <scope>test</scope>
                 </dependency>
                 <dependency>
                     <groupId>org.jboss.arquillian.junit</groupId>
                     <artifactId>arquillian-junit-container</artifactId>
                     <scope>test</scope>
-                    <version>1.4.1.Final</version>
+                    <version>${arquillian.version}</version>
                 </dependency>
                 <dependency>
                     <groupId>org.arquillian.container</groupId>
@@ -311,7 +311,7 @@
                 <dependency>
                     <groupId>org.jboss.arquillian.test</groupId>
                     <artifactId>arquillian-test-impl-base</artifactId>
-                    <version>1.4.1.Final</version>
+                    <version>${arquillian.version}</version>
                     <scope>test</scope>
                 </dependency>
 
@@ -319,13 +319,13 @@
                 <dependency>
                     <groupId>org.jboss.arquillian.core</groupId>
                     <artifactId>arquillian-core-impl-base</artifactId>
-                    <version>1.4.1.Final</version>
+                    <version>${arquillian.version}</version>
                     <scope>test</scope>
                 </dependency>
                 <dependency>
                     <groupId>org.jboss.arquillian.junit</groupId>
                     <artifactId>arquillian-junit-core</artifactId>
-                    <version>1.4.1.Final</version>
+                    <version>${arquillian.version}</version>
                     <scope>test</scope>
                 </dependency>
